refactor: Hint, Strike, Ball, Nothing Model 분리
Strike, Ball, Nothing 각각을 하나의 Model로 볼 수 있음.
