--- src/main/java/baseball/domain/Hint.java
@@ -1,32 +1,0 @@
-package baseball.domain;
-
-public class Hint {
-
-    private int ball;
-    private int strike;
-    private boolean nothing;
-
-    public int getBall() {
-        return ball;
-    }
-
-    public void setBall(int ball) {
-        this.ball = ball;
-    }
-
-    public int getStrike() {
-        return strike;
-    }
-
-    public void setStrike(int strike) {
-        this.strike = strike;
-    }
-
-    public boolean isNothing() {
-        return nothing;
-    }
-
-    public void setNothing(boolean nothing) {
-        this.nothing = nothing;
-    }
-}
