+++ src/main/java/baseball/domain/Hint/Ball.java
@@ -1,0 +1,7 @@
+package baseball.domain.Hint;
+
+public class Ball {
+
+    private int count;
+
+}
