+++ src/main/java/baseball/domain/Hint/Nothing.java
@@ -1,0 +1,7 @@
+package baseball.domain.Hint;
+
+public class Nothing {
+
+    private boolean status;
+
+}
