+++ src/main/java/baseball/domain/Hint/Strike.java
@@ -1,0 +1,7 @@
+package baseball.domain.Hint;
+
+public class Strike {
+
+    private int count = 0;
+
+}
